+++ appium-examples/src/test/java/com/realdevice/image_injection/Utils.java
@@ -1,4 +1,4 @@
-package com.realdevice.unifiedplatform.image_injection;
+package com.realdevice.image_injection;
 
 import io.appium.java_client.AppiumDriver;
 import io.appium.java_client.android.AndroidDriver;
