+++ appium-examples/src/test/java/com/realdevice/image_injection/images/qr-code.png
